+++ palmod/Game/Game_SFA3_A.cpp
@@ -506,7 +506,7 @@ BOOL CGame_SFA3_A::UpdatePalImg(int Node01, int Node02, int Node03, int Node04)
                 }
                 else
                 {
-                    UINT16 nDeltaToSecondElement = paletteDataSet->pPalettePairingInfo->nNodeIncrementToPartner;
+                    int8_t nDeltaToSecondElement = paletteDataSet->pPalettePairingInfo->nNodeIncrementToPartner;
 
                     fShouldUseAlternateLoadLogic = true;
                     UINT16 nPeerPaletteIdInNode = Node03 + nDeltaToSecondElement;
